# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agrisiti Academy API",
        "description": "REST and WebSocket API for the Agrisiti Academy e-learning platform. Covers student registration, course management, progress tracking, tests, assignments, notes, messaging, community forum, notifications, and the Agri chatbot (AI-powered onboarding assistant with real-time Reverb WebSocket streaming, SSE fallback, and course recommendations). See the Chatbot tag for the complete user journey, quiz questions, and 3-tier transport documentation.",
        "contact": {
            "email": "admin@agrisiti.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://academy-backends.agrisiti.com",
            "description": "Production Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/career/apprenticeships": {
            "get": {
                "tags": [
                    "Career Pathways"
                ],
                "summary": "Browse open apprenticeship slots (filterable by state, sector, and required course)",
                "operationId": "d12721b27be5e05875ef87e8df4bb14f",
                "parameters": [
                    {
                        "name": "state",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lga",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sector",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "required_course_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "show_all",
                        "in": "query",
                        "description": "If false (default), only slots whose required course the learner has completed (or that require no course) are shown",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated apprenticeship slots from approved organisations"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/career/apprenticeships/{slot}": {
            "get": {
                "tags": [
                    "Career Pathways"
                ],
                "summary": "Get a single apprenticeship slot's detail",
                "operationId": "f7db9858ce9e9571d9dba41817e66e71",
                "parameters": [
                    {
                        "name": "slot",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Slot detail"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/org/slots": {
            "get": {
                "tags": [
                    "Career Pathways"
                ],
                "summary": "List the authenticated organisation's own apprenticeship slots",
                "operationId": "2a1d5da8891d392e623a153dfc6655d5",
                "responses": {
                    "200": {
                        "description": "Organisation's slots with applicant counts"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Career Pathways"
                ],
                "summary": "Post a new apprenticeship slot (organisation must be admin-approved first)",
                "operationId": "ba6b5c06b534544b757a435f53736035",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "state",
                                    "duration"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "sector": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "state": {
                                        "type": "string"
                                    },
                                    "lga": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "duration": {
                                        "type": "string",
                                        "example": "3 months"
                                    },
                                    "required_course_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "openings": {
                                        "type": "integer",
                                        "default": 1
                                    },
                                    "application_deadline": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Slot created"
                    },
                    "403": {
                        "description": "Organisation is not yet approved by an admin"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/org/slots/{slot}": {
            "put": {
                "tags": [
                    "Career Pathways"
                ],
                "summary": "Update an apprenticeship slot the organisation owns",
                "operationId": "e070c03fbdc1e1c22071d604b8757135",
                "parameters": [
                    {
                        "name": "slot",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Slot updated"
                    },
                    "403": {
                        "description": "Not the owning organisation"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/org/slots/{slot}/close": {
            "post": {
                "tags": [
                    "Career Pathways"
                ],
                "summary": "Close an apprenticeship slot so it no longer appears in learner browsing",
                "operationId": "e15b92f86ffd7274e4a692d048af9088",
                "parameters": [
                    {
                        "name": "slot",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Slot closed"
                    },
                    "403": {
                        "description": "Not the owning organisation"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctumAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your Bearer token from /api/login"
            }
        }
    },
...
}